+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+  "trailingComma": "es5",
+  "tabWidth": 4,
+  "useTabs": true,
+  "semi": true,
+  "singleQuote": true,
+  "printWidth": 120
+}
