+++ src/app/station/components/station-resources/station-resources.component.html
