+++ src/app/station/components/station-summary/enums/resource-price-type.enum.ts
@@ -1,0 +1,7 @@
+/**
+ * The Resource Price Type enum
+ */
+export enum ResourcePriceType {
+	expense,
+	product,
+}
