+++ src/app/station/components/station-summary/interfaces/resource-amount.ts
@@ -1,0 +1,9 @@
+import { Ware } from 'src/app/shared/services/model/model';
+
+/**
+ * The Resource Amount interface
+ */
+export interface ResourceAmount {
+	ware: Ware;
+	value: number;
+}
