+++ src/app/station/components/station-summary/interfaces/resource-price-percentage.ts
@@ -1,0 +1,7 @@
+/**
+ * The Resource Price Percentage interface
+ */
+export interface ResourcePricePercentage {
+	resourcesPrice: number;
+	productsPrice: number;
+}
